Time: 112 ms (40.21%), Space: 14.4 MB (59.62%) - LeetHub

+++ 686-repeated-string-match/686-repeated-string-match.py
@@ -1,6 +1,6 @@
 class Solution:
     def repeatedStringMatch(self, a: str, b: str) -> int:
-        times = int(ceil(float(len(b)) / len(a)))
+        times = ceil(float(len(b)) / len(a))
         if b in a*times:
             return times
         if b in a*(times+1):
